pyproject.toml
[tool.poetry]
name = "weaviate-datahub"
version = "0.1.0"
description = ""
authors = ["Mousa Becquerel <m.sondoqah@becquerelinstitute.eu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandasai = ">=3.0.0b2"
gunicorn = "*"
flask = "*"
pydantic-ai = "*"
seaborn = "*"
pandasai-litellm = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-wtf = "*"
flask-login = "*"
flask-cors = "*"
flask-limiter = ">=3.6.0"
python-dotenv = "*"
weaviate-client = {extras = ["agents"], version = "*"}
psutil = "*"
logfire = {extras = ["flask"], version = "*"}
requests = "*"
beautifulsoup4 = "*"
psycopg = {extras = ["binary"], version = "^3.1.8"}
python-pptx = "^0.6.21"
openai-agents = {extras = ["redis", "sqlalchemy"], version = "^0.3.3"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"